Proper implementation for long-running loop
Including a significant clean-up of the code and its (debug) output.
Still a couple rough edges (mainly related to arguments and environment
passed in), but nearly there.

The provided delay between loops (hardcoded to 2,000 ms for now) is
kept as constant as possible (i.e. the time required to retrieve the
measurement is subtracted, so you should get a measurement every
"delay" amount of time).
The loop is asynchronous: The MQTT messages are send while we are
receiving a new batch of data from the TC66C. We'll only block on MQTT
if it's still sending when we have a new measurement available (which is
unlikely to happen under normal circumstances).
The minimum cycle-time is the maximum of the time required to receive
a measurement from the TC66C and the time required to send that
measurement to the MQTT broker. On my current test system (a RPi4) the
minimum cycle-time (based on the TC66C) is around 800 ms.

Most of the possible errors conditions should be handled gracefully.
They will still stop the script, but it will abort nicely (tearing down
all connections) and provide a sensible exit/status code. The remainder
of the handling can subsequently be done by the s6-overlay.