Further refinements; proper (configurable) logging
Add timestamped logging with configurable log-levels (debug, info, warn
and error) and a bit of colouring. The default log-level is "info",
which makes the standard output a bit more manageable.

Process all commandline parameters in a centralised place (still with
minimal validation though). Added (optional) "--interval" which can be
used to set (or disable) to measurement interval. Added (optional)
"--logLevel" which can be used to change the log-level.
The BLE MAC-address and MQTT broker can still be provided as positional
parameters (first and second respectively) or you can use "--bleAddress"
and "--mqttBroker" respectively.

Update README.md to match the latest changes and make the s6-overlay
run-script pass both "--interval" and "--logLevel" into the script
running into the container (if provided on the environment).

Finally, improved the way SIGINT is handled. We only capture the event
once, so pressing Ctrl+C multiple times (i.e. twice) will cause Node to
intercept and immediately terminate the script. Useful when the graceful
exit fails for whatever reason... Furthermore, reworked the "wait" logic
such that it responsive to SIGINT (instead of only responding at the end
of the wait cycle).