Long-running data retrieval (i.e. introduce a main-loop) (#1)
Instead of running once, the script now keeps running, fetching data at
a preset interval. On receiving SIGTERM or SIGINT it (attempts to)
gracefully terminate after finishing the current iteration. Failures should
lead to a non-zero exit-code.

The loop is asynchronous: The MQTT messages are send while we are
receiving a new batch of data from the TC66C. We'll only block on MQTT
if it's still sending when we have a new measurement available (which is
unlikely to happen under normal circumstances).
The minimum cycle-time is the maximum of the time required to receive
a measurement from the TC66C and the time required to send that
measurement to the MQTT broker. On my current test system (a RPi4) the
minimum cycle-time (based on the TC66C) is around 800 ms.

Most of the possible errors conditions should be handled gracefully.
They will still stop the script, but it will abort nicely (tearing down
all connections) and provide a sensible message and exit/status code.
The remainder of the handling can subsequently be done by the
s6-overlay.