Extend "Build and push" workflow
Instead of building and pushing to Docker Hub on pushes to main only, it
now builds on each push to GitHub (validating the Docker image along the
way). It still only pushes to Docker Hub on pushes to main.

The image is tagged not only as "latest", but also with the commit-ish
of the commit that triggered the build.

The push to Docker Hub is disabled at the moment; first want to validate
the workflow is doing what it's supposed to do...