Fix Dockerfile issues identified by hadolint
Pin all package versions to their current minor-release – as recommended
by Docker. Helps with (explicit) cache busting and reduces the
possibility of failures due to unexpected package updates.

Add "set -euxo pipefail" at the start of all RUN statements to ensure
they fail at the first sight of an error (and help with debugging
by echoing back the commands given).

Added the "--no-cache" back into the first "apk add" command. Not that
the caching matters (it happens in an ephemeral container), but it does
negate the need for "--update" (which was also missing) to ensure we get
the latest packages...