Docker and GitHub Actions tweaks (#2)
Instead of building and pushing to Docker Hub on pushes to main only, it
now builds on each push to GitHub (validating the Docker image along the
way). It still only pushes to Docker Hub on pushes to main.
The image is tagged not only as "latest", but also with the commit-ish
of the commit that triggered the build.

Also run Super-Linter on pushes to main. Otherwise the badge in README.md isn't
of much use (as it displays the state of the last run, instead of the state of
the last run on the main branch). Furthermore, enable VALIDATE_ALL_CODEBASE for
the main-branch only.

The above two changes are hard to test without actually merging this
pull-request to main – so here goes... 🚀

Use "npm ci" instead of "npm install" and add "--production" to ensure
dev-dependencies are not included. This reduces the size of the current
image by about 50 MiB (from 155 MiB to around 105 MiB).
Run the "npm ci" step in a separate image and only copy the resulting
artifacts into the actual image. This further reduces the size of the
image by around 10 MiB.

Pin all package versions in the Dockerfile to their current minor-release –
as recommended by Docker. Helps with (explicit) cache busting and reduces the
possibility of failures due to unexpected package updates.
Add "set -euo pipefail" at the start of all RUN statements to ensure
they fail at the first sight of an error 

Clean-up the s6-overlay helper scripts. Effectively combining the advise from
https://github.com/hadolint/hadolint/wiki/DL4006 with the advise from
https://github.com/koalaman/shellcheck/wiki/SC2086 to optimise the
scripts for use with /bin/ash inside Alpine.
They're now significantly more prone to failure on errors and instead of
disabling globbing I've implemented a solution with parameter expansion
as proposed by shellcheck.

Don't add UID/GID if the already exist. Docker-compose doesn't destroy the
container when (automatically) restarting it after a graceful shutdown. As such,
it's possible for the 10-useradd script to run inside a container that already
has the required UID/GID added (or the UID/GID could be present for other
reasons).