Update documentation
Somehow I missed the mark quite a bit in my initial investigation. There
was no fundamental problem with D-Bus; it appears to be a matter of
issuing a "command"-write instead of "reliable"-write...

Leaving the devices on my desk for a couple weeks and then trying again
quickly lead to the proper solution.