Support the 1.15 firmware/hardware revision (#3)
Automatically detects the correct characteristics to use. Furthermore, the
older firmware requires a reliable-write (whereas the newer firmware 
requires a command-write). The script now accounts for this and works
fine (and interchangeably) on both firmware/hardware revisions.

Update to node-ble 1.5.0 (which is required to be able to specify the
type of write to use).

Additionally, move all time-out handling into the main-loop. Decreases
complexity of the getDeviceCharacteristics() function. Also, an interrupt
signal will now always be acted upon right away (instead of only for half
of the blocking actions).

Finally, some Bluetooth-discovery related tweaks. Primarily to ensure we
don't accidentally leave discovery switched on if something goes wrong
while attempting to connect.