Add device alias to MQTT-topic (#4)
Allows multiple TC66C units to report to the same MQTT-broker. If no
deviceAlias is specified, the unit's Bluetooth MAC address is used.

Also, slightly reorganise the main flow: First connect to the TC66C and
only if that's successful, connect to the MQTT broker. Move more of
the main flow into the main try-catch (such that errors early on in the
process are also gracefully handled).

Furthermore, remove "set -u". This causes the shell to error if it
encounters undefined variables (which is not what we want: The script
is specifically set up to provide defaults for several optional - and thus
potentially undefined - variables).