What does this button do? 👩‍🔬
The Docker Hub "auto-tag" should now work as intended: Can't (easily)
tag images remotely; anyway, this appears to be a better approach.

Any push triggers a build of the Docker image. Any push to main causes
the image to be pushed to Docker Hub tagged with its commitish and
with "latest". Whenever a tag is created the image is pushed to Docker
Hub tagged with its commitish, the tag-name and with "stable" (ergo,
we assume a tag always indicates a stable release, something which could
be refined in the future as there's also pre-releases...).

Good enough for now. Also need to find a better way of testing GitHub
Actions. Pushing to main in an active repository isn't the quickest way
of figuring things out...