Major overhaul of the Docker workflow
Instead of unconditionally building the Docker image, only build it if
it isn't available on Docker Hub. Images are always tagged with their
Git commitish, which should ensure we only build each image once (and
prevent building/pushing multiple, functionally identical, copies of the
same image).

Additionally, revert my previous change with regards to tagging "stable"
and "latest". The "latest" tag should always (and only) be applied to
commits on the main-branch. It's not unthinkable that a release is not
created from the HEAD of main, but from an older commit. As such, it is
very well possible that when "stable" is created, it is not necessarily
"latest"...
Furthermore, only non-prerelease versions are tagged with "stable". The
others are only tagged with their version itself. Only "clean" semantic
versions (e.g. "1.2.3") are considered stable. If anything is appended
to the version (e.g. "1.2.3-beta2") it's considered prerelease.

Relying heavily on a (creative?) interpretation of the "job.if" feature
to ensure proper flow-control. Proof will be in the pudding...