Further refinements
Instead of failing the "pull"-job when the image isn't available on
Docker Hub, use output to communicate the result.

Also, buildx appears to not support storing multi-arch builds in a
tar-file, so for now we'll forgo storing a copy of the build as an
artifact of the workflow (marginally relevant to begin with).