Redo the Docker workflow
Using buildx to do multi-arch builds and passing around tar-files with
the resulting images doesn't as expected – the regular Docker-client
doesn't cope too well with the multi-arch aspect of it all...

So, it is either using a single step buildx-command to do everything
(building, tagging and pushing), or completely reimplement the
convenience of buildx using the regular build and the manifest-tool.

The latter would have allowed me to stay with a single, multi-step,
workflow with as little duplication in steps as possible. The former
appears much simpler though (even if it requires two workflows with a
lot of overlap).

So, there are now a "docker-ci" and a "docker-push" workflow, with some
overlap, but all-in-all a much clearer approach (and fewer magic
"if"-statements to control a highly dynamic flow).