GitHub Actions and Docker build process improvements (#5)
Using buildx to do multi-arch builds and passing around tar-files with
the resulting images doesn't work as expected – the regular Docker
client doesn't cope too well with the multi-arch aspect of it all...

So, it is either using a single step buildx-command to do everything
(building, tagging and pushing), or completely reimplement the
convenience of buildx using the regular build and the manifest-tool.

The latter would have allowed me to stay with a single, multi-step,
workflow with as little duplication in steps as possible. The former
appears much simpler though (even if it requires two workflows with a
lot of overlap).

So, there are now a "docker-ci" and a "docker-push" workflow, with some
overlap, but all-in-all a much clearer approach (and fewer magic
"if"-statements to control a highly dynamic flow).

Note that the "docker-push" workflow publishing a "named"-release
currently doesn't actually publish to Docker Hub; it just echo's out the
command it would use (as the only way to test this is to merge to main
and try it out).

Instruct JSCPD to ignore YAML-files. Yes, some of the workflows have
copy-pasted elements, there is no way around that...

Don't run the linter when a tag is created. Bit senseless as the tag
would always be created from a commit that has already been
linted... The previous behaviour leads to the GitHub interface
showing the linter check twice for commits that are tagged.

Also, remove "ci" environment from the linter. Has no added value,
just causes a lot of clutter in the GitHub interface.